requirements.txt
beautifulsoup4==4.12.3
bottle==0.13.2
certifi==2024.7.4
charset-normalizer==3.3.2
chat-downloader==0.2.8
colorama==0.4.6
colorlog==6.9.0
dnspython==2.7.0
docstring_parser==0.16
idna==3.7
isodate==0.7.2
pymongo==4.11.1
requests==2.32.3
soupsieve==2.5
urllib3==2.2.2
websocket-client==1.8.0
youtool==0.2.0
yt-dlp==2025.6.9
python-dotenv==1.0.1